+++ src/ais/mod.rs
@@ -31,6 +31,7 @@ pub(crate) mod vdm_t16;
 pub(crate) mod vdm_t17;
 pub(crate) mod vdm_t18;
 pub(crate) mod vdm_t19;
+pub(crate) mod vdm_t20;
 pub(crate) mod vdm_t21;
 pub(crate) mod vdm_t24;
 pub(crate) mod vdm_t27;
@@ -46,6 +47,7 @@ pub use vdm_t14::SafetyRelatedBroadcastMessage;
 pub use vdm_t15::{Interrogation, InterrogationCase};
 pub use vdm_t16::AssignmentModeCommand;
 pub use vdm_t17::DgnssBroadcastBinaryMessage;
+pub use vdm_t20::{DataLinkManagementMessage};
 pub use vdm_t21::{AidToNavigationReport, NavAidType};
 
 // -------------------------------------------------------------------------------------------------